Fix Synonym and SynonymOption Model CRUD option

+++ app/Http/Controllers/Teacher/Question/Vocabulary/Synonym/SynonymController.php
@@ -6,8 +6,12 @@ use App\Http\Controllers\Controller;
 use App\Model\Vocabulary\Synonym\Synonym;
 use App\Model\Vocabulary\Synonym\SynonymOption;
 use App\QuestionSet;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Crypt;
+use Illuminate\View\View;
 
 class SynonymController extends Controller
 {
@@ -16,7 +20,7 @@ class SynonymController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @return Factory|View
      */
     public function index()
     {
@@ -27,7 +31,7 @@ class SynonymController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @return Factory|View
      */
     public function create()
     {
@@ -39,28 +43,28 @@ class SynonymController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\RedirectResponse
+     * @param Request $request
+     * @return RedirectResponse
      */
     public function store(Request $request)
     {
-        $exam = $request->exam;
-        $set = $request->questionSet;
-
-        $countSynonymWordByExamAndSet = Synonym::where(['exam_id' => $exam, 'question_set_id' => $set])->get()->count();
+        $authTeacher = Auth::guard('teacher')->user();
+        $examId = $request->exam;
+        $setId = $request->questionSet;
+        $countSynonymWordByExamAndSet = $authTeacher->exams()->find($examId)->synonyms()->where('question_set_id', $setId)->get()->count();
 
         if ($countSynonymWordByExamAndSet < 5) {
 
             $synonymOptions = SynonymOption::create($this->validateSynonymOptionsCreateRequest($request));
             $this->synonymOptions = $synonymOptions;
 
-            Synonym::create($this->validateSynonymCreateRequest($request));
+            $authTeacher->exams()->find($examId)->synonyms()->create($this->validateSynonymCreateRequest($request));
 
             session()->flash('success_audio');
             toast('Synonym word has been successfully added','success');
         } else {
             session()->flash('field_audio');
-            alert()->info('Fail!', 'You can no longer add synonym word to this '. QuestionSet::find($set)->name .' set.');
+            alert()->info('Fail!', 'You can no longer add synonym word to this '. QuestionSet::find($setId)->name .' set.');
         }
         return redirect()->back();
     }
@@ -69,58 +73,73 @@ class SynonymController extends Controller
      * Display the specified resource.
      *
      * @param Synonym $synonym
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @return Factory|RedirectResponse|View
      */
     public function show(Synonym $synonym)
     {
-        return view('teacher.questions.vocabulary.synonym.show', compact('synonym'))
-            ->with('questionSets', QuestionSet::all())
-            ->with('authTeacher', Auth::guard('teacher')->user());
+        if ($this->validSynonymRequest($synonym)) {
+            return view('teacher.questions.vocabulary.synonym.show', compact('synonym'))
+                ->with('questionSets', QuestionSet::all())
+                ->with('authTeacher', Auth::guard('teacher')->user());
+        } else {
+            alert()->error('😒', 'You can\'t do this.');
+            return redirect()->back();
+        }
     }
 
     /**
      * Show the form for editing the specified resource.
      *
      * @param Synonym $synonym
-     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @return Factory|RedirectResponse|View
      */
     public function edit(Synonym $synonym)
     {
-        return view('teacher.questions.vocabulary.synonym.edit', compact('synonym'))
-            ->with('questionSets', QuestionSet::all())
-            ->with('authTeacher', Auth::guard('teacher')->user());
+        if ($this->validSynonymRequest($synonym)) {
+            return view('teacher.questions.vocabulary.synonym.edit', compact('synonym'))
+                ->with('questionSets', QuestionSet::all())
+                ->with('authTeacher', Auth::guard('teacher')->user());
+        } else {
+            alert()->error('😒', 'You can\'t do this.');
+            return redirect()->back();
+        }
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param \Illuminate\Http\Request $request
+     * @param Request $request
      * @param Synonym $synonym
-     * @return \Illuminate\Http\RedirectResponse
+     * @return RedirectResponse
      */
     public function update(Request $request, Synonym $synonym)
     {
-        $authTeacher = Auth::guard('teacher')->user();
-        $exam = $authTeacher->exams()->find($request->exam);
-        $set = $exam->sets()->find($request->questionSet);
+        if ($this->validSynonymRequest($synonym)) {
+            $authTeacher = Auth::guard('teacher')->user();
+            $exam = $authTeacher->exams()->find($request->exam);
+            $set = $exam->sets()->find($request->questionSet);
 
-        if ($synonym->exam->id == $exam->id && $synonym->set->id == $set->id) {
-            $synonym->update($this->validateSynonymUpdateRequest($request));
-            session()->flash('success_audio');
-            toast('Synonym Word has been successfully updated','success');
-            return redirect()->route('teachers.questions.synonyms.show', $synonym->id);
-        }
+            if ($synonym->exam->id == $exam->id && $synonym->set->id == $set->id) {
+                $synonym->update($this->validateSynonymUpdateRequest($request));
+                session()->flash('success_audio');
+                toast('Synonym Word has been successfully updated','success');
+                return redirect(route('teachers.questions.synonyms.show', $synonym->id).'?exam='.request()->get('exam').'&set='.request()->get('set'));
+            }
 
-        $countSynonymWordByExamAndSet = Synonym::where(['exam_id' => $exam->id, 'question_set_id' => $set->id])->get()->count();
+            $countSynonymWordByExamAndSet = Synonym::where(['exam_id' => $exam->id, 'question_set_id' => $set->id])->get()->count();
 
-        if ($countSynonymWordByExamAndSet < 5) {
-            $synonym->update($this->validateSynonymUpdateRequest($request));
-            session()->flash('success_audio');
-            toast('Synonym Word has been successfully updated','success');
-            return redirect()->route('teachers.questions.synonyms.show', $synonym->id);
+            if ($countSynonymWordByExamAndSet < 5) {
+                $synonym->update($this->validateSynonymUpdateRequest($request));
+                session()->flash('success_audio');
+                toast('Synonym Word has been successfully updated','success');
+                return redirect(route('teachers.questions.synonyms.show', $synonym->id).'?exam='.request()->get('exam').'&set='.request()->get('set'));
+            } else {
+                session()->flash('field_audio');
+                alert()->info('Fail!', 'You can no longer add synonym word to this '. QuestionSet::find($set->id)->name .' set.');
+                return redirect()->back();
+            }
         } else {
-            session()->flash('field_audio');
-            alert()->info('Fail!', 'You can no longer add synonym word to this '. QuestionSet::find($set->id)->name .' set.');
+            alert()->error('😒', 'You can\'t do this.');
             return redirect()->back();
         }
     }
@@ -129,14 +148,21 @@ class SynonymController extends Controller
      * Remove the specified resource from storage.
      *
      * @param Synonym $synonym
-     * @return \Illuminate\Http\RedirectResponse
+     * @return RedirectResponse
      */
     public function destroy(Synonym $synonym)
     {
-        $synonym->forceDelete();
-        session()->flash('success_audio');
-        toast('Synonym word has been successfully deleted','success');
-        return redirect()->route('teachers.questions.synonyms.index');
+        if ($this->validSynonymRequest($synonym)) {
+            $synonym->answer()->forceDelete();
+
+            $synonym->forceDelete();
+            session()->flash('success_audio');
+            toast('Synonym word has been successfully deleted','success');
+            return redirect()->route('teachers.questions.synonyms.index');
+        } else {
+            alert()->error('😒', 'You can\'t do this.');
+            return redirect()->back();
+        }
     }
 
     private function validateSynonymOptionsCreateRequest($request)
@@ -155,17 +181,31 @@ class SynonymController extends Controller
 
     }
 
+    private function validSynonymRequest($synonym) {
+
+        $examId = Crypt::decrypt(\request()->get('exam'));
+        $setId = Crypt::decrypt(\request()->get('set'));
+        $authTeacherOptionByExamAndSet = Auth::guard('teacher')->user()->exams()->find($examId)->synonyms()->where('question_set_id', $setId)->get();
+
+        $valid = null;
+        foreach ($authTeacherOptionByExamAndSet as $item) {
+            if ($item->id === $synonym->id) {
+                $valid = true;
+            }
+        }
+
+        return $valid;
+    }
+
     private function validateSynonymCreateRequest($request)
     {
         $validateData = $this->validate($request, [
-            'exam' => 'required|integer',
             'questionSet' => 'required|integer',
             'word' => 'required|string|max:255',
             'answer' => 'required|string|max:255',
         ]);
 
         return [
-            'exam_id' => $validateData['exam'],
             'question_set_id' => $validateData['questionSet'],
             'synonym_option_id' => $this->synonymOptions['id'],
             'word' => $validateData['word'],
@@ -188,5 +228,6 @@ class SynonymController extends Controller
         ];
 
     }
+
 
 }
